Объединения формы создание и редактирования. Переделал генерацию фото
