+++ src/mock/event.js
@@ -15,7 +15,11 @@ const typeRoutes = [
   { title: 'Train', img: 'img/icons/train.png', allOffer: [], selectedOffer: [], allPriceOffers: 0 }
 ];
 
-const cityes = ['Amsterdam', 'Geneva', 'Chamonix'];
+const cityes = [
+  { titleCity: 'Amsterdam', description: '', photos: [] },
+  { titleCity: 'Geneva', description: '', photos: [] },
+  { titleCity: 'Chamonix', description: '', photos: [] }
+];
 
 const offersTitle = [
   { 'text': 'Add luggage', 'id': 'luggage' },
@@ -42,8 +46,6 @@ const generateDate = () => {
 };
 
 const generateTime = (date) => {
-  const dateBegin = new Date('', dayjs(date.dataBeginEvent).format('M'), dayjs(date.dataBeginEvent).format('D'), dayjs(date.dataBeginEvent).format('H'), dayjs(date.dataBeginEvent).format('m'));
-  const dateEnd = new Date('', dayjs(date.dataEndEvent).format('M'), dayjs(date.dataEndEvent).format('D'), dayjs(date.dataEndEvent).format('H'), dayjs(date.dataEndEvent).format('m'));
   const duration = getDiffDates(date.dataBeginEvent, date.dataEndEvent);
   let durationFormat = '';
   if (duration.days !== 0) {
@@ -65,17 +67,16 @@ const generateTime = (date) => {
 };
 
 const generateDescription = () => {
-  const descriptionArray = 'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Cras aliquet varius magna, non porta ligula feugiat eget. Fusce tristique felis at fermentum pharetra. Aliquam id orci ut lectus varius viverra. Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante. Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum. Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui. Sed sed nisi sed augue convallis suscipit in sed felis. Aliquam erat volutpat. Nunc fermentum tortor ac porta dapibus. In rutrum ac purus sit amet tempus.'.split('.');
-  const countDescription = getRandomNumber(1, descriptionArray.length);
-  let description = '';
-  for (let i = 0; i < countDescription; i++) {
-    const elementNumber = getRandomNumber(0, descriptionArray.length - 1);
-    const descriptionArrayElement = descriptionArray[elementNumber];
-    descriptionArray.splice(elementNumber, 1);
-    description += descriptionArrayElement;
-  }
-
-  return description;
+  cityes.forEach((city) => {
+    const descriptionArray = 'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Cras aliquet varius magna, non porta ligula feugiat eget. Fusce tristique felis at fermentum pharetra. Aliquam id orci ut lectus varius viverra. Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante. Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum. Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui. Sed sed nisi sed augue convallis suscipit in sed felis. Aliquam erat volutpat. Nunc fermentum tortor ac porta dapibus. In rutrum ac purus sit amet tempus.'.split('.');
+    const countDescription = getRandomNumber(1, descriptionArray.length);
+    for (let i = 0; i < countDescription; i++) {
+      const elementNumber = getRandomNumber(0, descriptionArray.length - 1);
+      const descriptionArrayElement = descriptionArray[elementNumber];
+      descriptionArray.splice(elementNumber, 1);
+      city.description += descriptionArrayElement;
+    }
+  })
 };
 
 const generateOffers = () => {
@@ -99,22 +100,22 @@ const generateOffers = () => {
       }
     }
   });
-  console.log(typeRoutes);
+};
+
+const generatePhoto = () => {
+  cityes.forEach((city) => {
+    let numberPhoto = 0;
+    const countPhotos = getRandomNumber(3, 6);
+    for (let i = 0; i < countPhotos; i++) {
+      numberPhoto += getRandomNumber(1, 10);
+      city.photos.push(`http://picsum.photos/248/152?r=${numberPhoto}`);
+    }
+  })
 };
 
 generateOffers();
-
-const generatePhoto = () => {
-  const photos = [];
-  let numberPhoto = 0;
-  const countPhotos = getRandomNumber(3, 6);
-  for (let i = 0; i < countPhotos; i++) {
-    numberPhoto += getRandomNumber(1, 10);
-    photos.push(`http://picsum.photos/248/152?r=${numberPhoto}`);
-  }
-
-  return photos;
-};
+generateDescription();
+generatePhoto();
 
 const generateEvents = () => {
   const date = generateDate();
@@ -127,10 +128,8 @@ const generateEvents = () => {
     type,
     city: cityes[getRandomNumber(0, 2)],
     time,
-    description: generateDescription(),
     allPrice,
     favorite: Boolean(getRandomNumber(0, 1)),
-    photos: generatePhoto()
   };
 };
