+++ src/view/item-event-view.js
@@ -20,7 +20,7 @@ const createTripEvent = (event) => {
     <div class="event__type">
       <img class="event__type-icon" width="42" height="42" src="${img}" alt="Event type icon">
     </div>
-    <h3 class="event__title">${title} ${city}</h3>
+    <h3 class="event__title">${title} ${city.titleCity}</h3>
     <div class="event__schedule">
       <p class="event__time">
         <time class="event__start-time" datetime="2019-03-18T10:30">${time.startTime}</time>
